refactor(usm): introduce EngineBoots and EngineTime newtypes
Step 1 of WI-2026-06-12-001. New src/usm/engine_time.rs defines
EngineBoots(u32) and EngineTime(u32); is_in_time_window now
distinguishes boots from time at the type level, removing the
four-indistinguishable-u32s transposition risk in RFC 3414
replay-window validation. EngineBoots owns the RFC 3414 ceiling:
MAX_ENGINE_BOOTS delegates to EngineBoots::CEILING, and saturating
increment / is_at_ceiling behaviour is unit-tested in the newtype,
including above-ceiling inputs.

The newtypes thread through DispatchInputs, EventLoop, TrapSender,
dispatch/report call sites, and AgentBuilder::build. Public API is
unchanged: EngineBootsStore, initialise_engine_boots,
UsmSecurityFields, and encode_v3_* keep their u32 signatures with
conversions at those boundaries. The engine_time module is pub +
doc(hidden) for fuzz access, following the dispatch_context
precedent; fuzz targets updated.

Gates: clippy pedantic clean, 685 unit tests, make pre-commit,
51/51 behave scenarios.

+++ src/transport/dispatch/report.rs
@@ -108,8 +108,8 @@ pub(super) fn emit_report_response(
     crate::codec::encode_v3_report(
         msg_id,
         ctx.inputs.engine_id,
-        ctx.inputs.engine_boots,
-        ctx.inputs.engine_time,
+        u32::from(ctx.inputs.engine_boots),
+        u32::from(ctx.inputs.engine_time),
         counter_oid,
         counter_value,
     )
